+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Cover `<MappedHeader>` component with unit tests (UIDATIMP-693)
 * Reuse `<JobsListAccordion>` component from `stripes-data-transfer-components` rep (UIDATIMP-574)
 * Reuse `<JobsLogs>` component from `stripes-data-transfer-components` rep (UIDATIMP-572)
+* Reuse `<Preloader>` component from `stripes-data-transfer-components` rep (UIDATIMP-580)
 
 ### Bugs fixed:
 * Log filter dropdowns by job profile and user are not in alphabetical order (UIDATIMP-601)